+++ earth2mip/regrid.py
@@ -18,6 +18,8 @@ import netCDF4 as nc
 import einops
 import torch
 import pathlib
+import xarray
+import numpy as np
 from earth2mip.schema import Grid
 from earth2mip._config import Settings
 
@@ -57,6 +59,11 @@ class Identity(torch.nn.Module):
         return x
 
 
+class DropSouthPole(torch.nn.Module):
+    def forward(self, x):
+        return x[..., :-1, :]
+
+
 def _get_tempest_regridder(src: Grid, dest: Grid) -> TempestRegridder:
     # TODO map data needs to be available for S2S scoring
     config = Settings()
@@ -71,6 +78,34 @@ def _get_tempest_regridder(src: Grid, dest: Grid) -> TempestRegridder:
 def get_regridder(src: Grid, dest: Grid):
     if src == dest:
         return Identity()
+    elif (src, dest) == (Grid.grid_721x1440, Grid.grid_720x1440):
+        return DropSouthPole()
     else:
         return _get_tempest_regridder(src, dest)
     raise NotImplementedError()
+
+
+def xarray_regrid(src: xarray.Dataset, dest: Grid):
+    """Simple function for regridding an xarray dataset to a grid.
+    The xarray dataset must have lat and lon coords
+
+    Parameters
+    ----------
+    src : xr.Dataset
+        Input xarray data set
+    dest : Grid
+        Target grid scheme
+    """
+    # TODO refactor this into a regridder
+    # Subsample / interpolate lat lon grid
+    if np.isin(dest.lat, src.coords["lat"].values).all():
+        src = src.sel(lat=dest.lat)
+    else:
+        src = src.interp(lat=dest.lat)
+
+    if np.isin(dest.lon, src.coords["lon"].values).all():
+        src = src.sel(lon=dest.lon)
+    else:
+        src = src.interp(lon=dest.lon)
+
+    return src