+++ tests/test_end_to_end.py
@@ -19,7 +19,8 @@ import torch
 import xarray
 import hashlib
 import numpy as np
-from earth2mip.networks import persistence
+from earth2mip.inference_ensemble import run_basic_inference
+from earth2mip.networks import persistence, get_model
 from earth2mip import (
     schema,
     weather_events,
@@ -28,6 +29,7 @@ from earth2mip import (
     inference_medium_range,
 )
 import pytest
+from earth2mip._channel_stds import channel_stds
 
 
 def checksum_reduce_precision(arr, digits=3):
@@ -39,25 +41,25 @@ def checksum_reduce_precision(arr, digits=3):
 
 
 class get_data_source:
+    grid = schema.Grid.grid_721x1440
+
     def __init__(self, inference):
-        arr = xarray.DataArray(
-            np.ones([1, len(inference.in_channel_names), 721, 1440]),
-            dims=["time", "channel", "lat", "lon"],
-        )
-        arr["channel"] = inference.in_channel_names
-        arr["lat"] = np.linspace(90, -90, 721)
-        arr["lon"] = np.linspace(0, 360, 1440, endpoint=False)
+        self.channel_names = inference.in_channel_names
+        arr = np.ones([1, len(inference.in_channel_names), 721, 1440])
         self.arr = arr
         self.channel_names = inference.out_channel_names
 
-    def __getitem__(self, time):
-        return self.arr.assign_coords(time=[time])
+    def __getitem__(self, time) -> np.ndarray:
+        return self.arr
 
 
 def test_inference_ensemble(tmp_path):
     inference = persistence(package=None)
     data_source = get_data_source(inference)
     time = datetime.datetime(2018, 1, 1)
+    channel_stds["a"] = 0.0
+    channel_stds["b"] = 0.0
+    channel_stds["c"] = 0.0
     config = schema.EnsembleRun(
         weather_model="dummy",
         simulation_length=40,
@@ -125,3 +127,13 @@ def test_inference_medium_range(tmpdir, regtest):
         metrics.rmse, digits=3
     )
     metrics.info(regtest)
+
+
+@pytest.mark.slow
+def test_run_basic_inference():
+    time_loop = get_model("e2mip://fcn", device="cuda:0")
+    data_source = get_data_source(time_loop)
+    ds = run_basic_inference(
+        time_loop, n=10, data_source=data_source, time=datetime.datetime(2018, 1, 1)
+    )
+    print(ds)