+++ CHANGELOG.md
@@ -15,6 +15,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- `earth2mip.schema.ChannelSet` removed and
+`earth2mip.schema.Model.in_channel_names` added. **Some older model packages
+will be broken and can be fixed by adding .in_channel_names and
+.out_channel_names attributes to any metadata.json files**.
+- The `ERA5_HDF_73` and `TIME_MEAN_73` configurations have been removed. Used
+`ERA5_HDF` and `TIME_MEAN` instead to control the data used by the "era5" or
+"hrmip" initial condition sources.
 - Add `--model-metadata` flag to `earth2mip.inference_medium_range` and
   `earth2mip.inference_ensemble`.
 - Add `metadata` argument to `earth2mip.networks.get_model`.